test(backup): keep CLI proof in process shard
